add convert to xmp script

+++ dump/mage.py
@@ -6,6 +6,8 @@ import datetime
 from json import JSONEncoder
 import re
 import gzip
+from convert_to_xmp import create_xmp_metadata
+
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
@@ -64,22 +66,23 @@ def extract_braces(html_content):
 
 
 def yield_images():
-    cookie = "ph_phc_lBVvvz084lS4XFbPqGf38TGEa6HnGhmgVHcf4f0NNuX_posthog=%7B%22distinct_id%22%3A%22dgGk4FOcMSf5KbIBqkQAteJolLB3%22%2C%22%24sesid%22%3A%5B1741642694000%2C%22019581fd-98b7-7fda-a4d0-547ed132a3bb%22%2C1741642635447%5D%2C%22%24epp%22%3Atrue%7D; __session=eyJhbGciOiJSUzI1NiIsImtpZCI6IlFEZ1JhQSJ9.eyJpc3MiOiJodHRwczovL3Nlc3Npb24uZmlyZWJhc2UuZ29vZ2xlLmNvbS9tYWdlZG90c3BhY2UiLCJuYW1lIjoiSmVmZnJleSBNZWxsb3kiLCJwaWN0dXJlIjoiaHR0cHM6Ly9saDMuZ29vZ2xldXNlcmNvbnRlbnQuY29tL2EvQUNnOG9jSkV1TjNHamtHRWZucG9zNFlXcVhkdHB6cUxTbi0wb3dZZUpDVmhZSURpbEFcdTAwM2RzOTYtYyIsInN0cmlwZVJvbGUiOiJwcm8iLCJhdWQiOiJtYWdlZG90c3BhY2UiLCJhdXRoX3RpbWUiOjE3Mzk4OTk1NDcsInVzZXJfaWQiOiJkZ0drNEZPY01TZjVLYklCcWtRQXRlSm9sTEIzIiwic3ViIjoiZGdHazRGT2NNU2Y1S2JJQnFrUUF0ZUpvbExCMyIsImlhdCI6MTc0MTY0MjYzNywiZXhwIjoxNzQxNjU3MDM3LCJlbWFpbCI6ImptZWxsb3lAZ21haWwuY29tIiwiZW1haWxfdmVyaWZpZWQiOnRydWUsImZpcmViYXNlIjp7ImlkZW50aXRpZXMiOnsiZ29vZ2xlLmNvbSI6WyIxMDM1NDE3NTI5NjM2MjY2NDg0MzAiXSwiZW1haWwiOlsiam1lbGxveUBnbWFpbC5jb20iXX0sInNpZ25faW5fcHJvdmlkZXIiOiJnb29nbGUuY29tIn19.lRS7jSu20jHDEXHgo365tRoZzDV0CUGTNY5Yrez1xQP22FZ_uDUT0HgD-KyBh205_tzE0D4X-fuaUbL7LTxb5mFF_QpzlgTPJexOmJXpW5hx8MWiHzKLgH-HTuqljaG3S3qNsb-rdXzsD8Dh-12fXxk6EOaRME8XEBlyFV2LN2VHqaCqJwCObzdfayr04V3STv83et-mmuEKw6kNmBxM6i2cqzGIZXq8nhgQEYw-8sGu7wyFZ7izp8FeCcj1IJqGsk2rErDxV9ivbkN1UgF8rxhN-4_Rw0z6yN6pSe_A2_GlcYcSx-4HPjsMtfsHtygUT6e_xkOhxeY3emurUhsY7Q; cf_clearance=qWeKPd2SnxXWauD4cPf_aiBQ6Xwk5tRxDvWgqbeVSds-1741642635-1.2.1.1-3xtCur8HosX55gRItNUepZAp4yOWKggtBNgEB7lUiDfaFCyizVeQN2qtmwHdo4dJyNHu2uW7JS9rkYJeFKoT679HsoPDKvrPeZ4fz6mWp8rGc2RIEcxMhCLRl2iRCjPAawUeAGRljDKp_S46WklhZdMq57.aoAPish8htPJ8n1lo4e68Pu7lznQMTgYZ8y2zGMs9aFfd5_feN1jzZY80VAcv0n4GaHBSAk5fPKCfzv7KIjBrQ9WF7.bJcS9DwsvAzbKnEBiI_31RQ_Ehvk9RsyqOVoeJt9hAze_99wmm1BNYE1liBy0kn_RjNJrRhfe.Cv_fk19fMDToTWGRRSSEjXqAFNBxCHsmbbc6Ic3e3R0; _iidt=JU+/c/irOfY2QalLMMMN+4IodD5VfEq4VP+/6KYwgAW/Jy29tHaA5huTpwu15tVXoTQ/C8ZUF+zltg==; _vid_t=RPaHoBvynXjy3c/wnHnRw/W6hHS+lhCl1fgE4rQDVoUP/SAwTPhc5x7kncI3HBXDDe9UHPmgybM+iA=="
+    cookie = "ph_phc_lBVvvz084lS4XFbPqGf38TGEa6HnGhmgVHcf4f0NNuX_posthog=%7B%22distinct_id%22%3A%22dgGk4FOcMSf5KbIBqkQAteJolLB3%22%2C%22%24sesid%22%3A%5B1748822103918%2C%2201972de8-9dca-7ac0-a78a-59c37dfc5db9%22%2C1748821908938%5D%2C%22%24epp%22%3Atrue%2C%22%24initial_person_info%22%3A%7B%22r%22%3A%22%24direct%22%2C%22u%22%3A%22https%3A%2F%2Fwww.mage.space%2F%22%7D%7D; __session=eyJhbGciOiJSUzI1NiIsImtpZCI6Ik4yQlkwQSJ9.eyJpc3MiOiJodHRwczovL3Nlc3Npb24uZmlyZWJhc2UuZ29vZ2xlLmNvbS9tYWdlZG90c3BhY2UiLCJuYW1lIjoiSmVmZnJleSBNZWxsb3kiLCJwaWN0dXJlIjoiaHR0cHM6Ly9saDMuZ29vZ2xldXNlcmNvbnRlbnQuY29tL2EvQUNnOG9jSkV1TjNHamtHRWZucG9zNFlXcVhkdHB6cUxTbi0wb3dZZUpDVmhZSURpbEFcdTAwM2RzOTYtYyIsInN0cmlwZVJvbGUiOiJwcm9fcGx1cyIsImF1ZCI6Im1hZ2Vkb3RzcGFjZSIsImF1dGhfdGltZSI6MTc0NzE2NjcwNCwidXNlcl9pZCI6ImRnR2s0Rk9jTVNmNUtiSUJxa1FBdGVKb2xMQjMiLCJzdWIiOiJkZ0drNEZPY01TZjVLYklCcWtRQXRlSm9sTEIzIiwiaWF0IjoxNzQ4ODIyMTAxLCJleHAiOjE3NDg5MDg1MDEsImVtYWlsIjoiam1lbGxveUBnbWFpbC5jb20iLCJlbWFpbF92ZXJpZmllZCI6dHJ1ZSwiZmlyZWJhc2UiOnsiaWRlbnRpdGllcyI6eyJnb29nbGUuY29tIjpbIjEwMzU0MTc1Mjk2MzYyNjY0ODQzMCJdLCJlbWFpbCI6WyJqbWVsbG95QGdtYWlsLmNvbSJdfSwic2lnbl9pbl9wcm92aWRlciI6Imdvb2dsZS5jb20ifX0.NR7X8VcLXkYlkz4ltW-yD99rWxinFW9oIOdZZK-kYUxn_y8ScpKuXIqUsn49WGIzacLXGtRUJrBrMdaKg7m3AvOAkA8OV8ljosNb6DDycEnzdp_TMl4-iyf_05QNKxdzCrBMLYJTD7JiiG9KVtIxzVGrYy9DOG3377YEoGvaID3HO8JPQYv7QaROFZVsgAzr01JBKbl7V8Oxp_FeuSzqFyjvAbQ7hIqkto--3Epmn_vpVy9rxJJeIoheP7C-fVIJqNaoaXprxA28JeRgAUxmZdnU3-YQPXbLCaQeB-pkdC4h3RJFzigFDBSkZ0mtiIAZuBN83In9zkGiXVPDM5Te0A; cf_clearance=sdjQDJRnIWpVTwMzq51dlsZydDFgAnTMqUI7Ay9vXkg-1748821909-1.2.1.1-m55SVLynj9WmQBgRlhW0odJ5XVjo_sjdmRsYjgUXyKiq7hLvoI_SMlY4SEUYfcBKni9xCfyZ5zQsA3y_k8YmEDjq8RUErDcV_PaWhtzdmvKcEEj5bivtawse7Q52ZOYs7f7qMGdDlAtynWWhMNLqkqjdxpag61m9zAQF3onO51hj6one_W8HalqVVs5n2yoHJOLTZ2TWB5ejlyR_QLiXFvUKKlkE.pNwZLq3JWcNAUT0qPYmW.s7uOZ2Sc3AF3czLRQFHfocxKK8gUTsZeaNXzGx6O3Tv.Zf3dIMavnkPRmEsWRBB8Sj_k.6a2AV.Lngug2MKDtmEwjVM4MdfkqGWpBKNjHRFX19xyVqYvH055s; _iidt=EU88TS7T3Nqc25zngk6fq+6Dt9ScFjYSsR7cjMB8+GvGi8NY5HFftA4dbq6HOj+/GoKY9EPuSXynSydN0rnBP4GVyKGmPFk3be1vBnk="
 
-    request = [
-        "dgGk4FOcMSf5KbIBqkQAteJolLB3",
-        100,
-        0,
-        {"orderBy": "desc", "prompt": "$undefined", "type": "$undefined"},
-    ]
+    request = ["dgGk4FOcMSf5KbIBqkQAteJolLB3", 100, 0, {
+        "orderBy": "desc",
+        "prompt": "$undefined",
+        "type": "$undefined",
+        "tag": "$undefined"
+    }]
 
-    r = requests.Session()
-    r.cookies.update(
+
+    session = requests.Session()
+    session.cookies.update(
         {
             "cookie": cookie,
         }
     )
-    r.headers.update(
+    session.headers.update(
         {
             "content-type": "text/plain",
             "referer": "https://www.mage.space/u/jmelloy",
@@ -95,14 +98,16 @@ def yield_images():
         seen = False
         request[2] = offset
         print(request)
-        text, _ = wrap(r, url, data=request)
+        text, _ = wrap(session, url, data=json.dumps(request))
 
-        potential_json = extract_braces(text)
+        potential_json = extract_braces(text.decode("utf-8"))
         for json_str in sorted(potential_json, key=len, reverse=True):
             if seen:
                 continue
             try:
+                json_str = re.sub(r'"embedding":\S+\[.*?\],?', '', json_str.replace('\\"', '"')).replace(",}", "}")
                 data = json.loads(json_str.replace('\\"', '"'))
+                
 
                 if "uid" in data and "children" in data:
                     images = data["children"][-1]["creations"]
@@ -127,20 +132,27 @@ def yeild_from_file():
     
     for row in extract_braces(r):
         data = json.loads(row)
-
+        if "creations" not in data:
+            continue
         for img in data["creations"]:
-            yield img
+            if img:
+                yield img
             
 def main():
     seen = set()
     for image in yeild_from_file():
         if image["id"] in seen:
-            raise Exception("Already seen")
+            continue
+            # raise Exception("Already seen")
         seen.add(image["id"])
+        prompt = (image.get("concept_override") or {}).get("prompt", "")
+        if not prompt:
+            prompt = (image.get("architecture_config") or {}).get("prompt", "")
+
         print(
             image["created_at"],
             image["url"].split("/")[-1],
-            image.get("concept_override", {}).get("prompt"),
+            prompt
         )
 
         filename = os.path.join(
@@ -155,14 +167,17 @@ def main():
         if not os.path.exists(os.path.dirname(filename)):
             os.makedirs(os.path.dirname(filename), exist_ok=True)
 
-        with open(filename, "wb") as f:
-            f.write(requests.get(image["url"]).content)
+        if not os.path.exists(filename):
+            with open(filename, "wb") as f:
+                f.write(requests.get(image["url"]).content)
 
         with open(os.path.join(os.path.dirname(filename), "meta.json"), "w") as f:
             f.write(json.dumps(image, indent=2))
 
-        seen.add(image["id"])
 
+        sidecar = create_xmp_metadata(image)
+        with open(f"{filename}.xmp", "w") as f:
+            f.write(sidecar)
 
 if __name__ == "__main__":
     main()
